[WIP] Added deterministic module
